moved test related client internal stuff to its own tag
